!2 v1.0 * 1、增加打包脚本

+++ os/src/cstream.cpp
@@ -186,7 +186,7 @@ CStream &CStream::operator<<(CStream &t)
 }
 WORD CStream::to_WORD()
 {
-    return to_DWORD();
+    return (WORD)to_DWORD();
 }
 DWORD CStream::to_DWORD()
 {
